Validator class reuse Validator class moved to Util package Unit Test cases for Validator class
